your commit message

+++ src/heredoc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   heredoc.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ksevciko <ksevciko@student.42prague.com    +#+  +:+       +#+        */
+/*   By: jasminelager <jasminelager@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/30 12:34:17 by ksevciko          #+#    #+#             */
-/*   Updated: 2025/10/03 00:13:15 by ksevciko         ###   ########.fr       */
+/*   Updated: 2025/10/08 10:35:13 by jasminelage      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ bool	rm_quotes_delim(t_mini *var, char **delim)
 	len = len_no_quotes(*delim, &quoted);
 	result = (char *)malloc((len + 1) * sizeof(char));
 	if (!result)
-		error_exit(var, BOLD RED "minishell: malloc failed\n" RESET);
+		error_exit(var, "minishell: malloc failed\n");
 	cpy_expanded_delim(*delim, result);
 	free(*delim);
 	*delim = result;
@@ -62,7 +62,7 @@ void	read_heredoc(t_mini *var, char *delim, int fd, bool delim_quoted)
 {
 	char	*line_in;
 
-	line_in = readline(BOLD GRAY ">" RESET);
+	line_in = readline("> ");
 	while (ft_strncmp(line_in, delim, ft_strlen(delim) + 1))
 	{
 		if (!line_in)
@@ -72,7 +72,7 @@ void	read_heredoc(t_mini *var, char *delim, int fd, bool delim_quoted)
 		write(fd, line_in, ft_strlen(line_in));
 		write(fd, "\n", 1);
 		free(line_in);
-		line_in = readline(BOLD GRAY ">" RESET);
+		line_in = readline("> ");
 	}
 	free(line_in);
 }
@@ -82,6 +82,7 @@ void	heredoc_to_file(t_mini *var, char *filename, char **delim)
 	int		fd;
 	bool	delim_quoted;
 
+	signals_heredoc();  // Set heredoc signal handler
 	fd = open_tmp_file(var, filename);
 	delim_quoted = rm_quotes_delim(var, delim);
 	read_heredoc(var, *delim, fd, delim_quoted);
@@ -103,16 +104,21 @@ bool	heredoc(t_mini *var, t_token *delim)
 	if (pid == -1)
 	{
 		perror("fork");
-		return (0); //make it return all the way to main
+		free(file);
+		return (0);
 	}
 	else if (pid == 0)
 		heredoc_to_file(var, file, &delim->content);
 	waitpid(pid, &status, 0);
-	if (WIFSIGNALED(status)) // kinda what a signal should do here (AI generated, please check it)
+	signals_setup();  // Restore interactive signal handlers after heredoc completes
+	if (WIFSIGNALED(status) && WTERMSIG(status) == SIGINT)
 	{
 		unlink(file);
-		g_signal = 130;  // 128 + SIGINT
-		return (0);  // return all the way to main immediately, do not execute anything
+		free(file);
+		var->exit_code = 130;
+		g_signal = 130;  // Set global signal flag
+		signals_setup();  // Restore interactive signal handlers
+		return (0);
 	}
 	free(delim->content);
 	delim->content = file;
